Split testing projects. - Divided the server and client library testing classes into their own respective projects. This should make testing easier when implementing a change in different parts of the solution.
