+++ ClientTests/ClientTests.csproj
