+++ NUTDotNet.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 16.0.30907.101
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NUTDotNetClient", "NUTDotNetClient\NUTDotNetClient.csproj", "{536AB832-D67B-46DD-B748-BF144A2DAD29}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Testing", "Testing\Testing.csproj", "{7F3DFA8A-7EEE-4A50-A23A-9F3E77D5897A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ClientTests", "ClientTests\ClientTests.csproj", "{7F3DFA8A-7EEE-4A50-A23A-9F3E77D5897A}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{34CC86D2-CE11-4A70-A99F-E9BE8C9A3CED}"
 	ProjectSection(SolutionItems) = preProject
@@ -14,7 +14,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NUTDotNetServer", "NUTDotNetServer\NUTDotNetServer.csproj", "{6F262F4A-8FD4-44DB-A6BE-52340B75ADB7}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NUTDotNetShared", "NUTDotNetShared\NUTDotNetShared.csproj", "{E49D61BD-AC6D-492E-B098-190166F0C44B}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NUTDotNetShared", "NUTDotNetShared\NUTDotNetShared.csproj", "{E49D61BD-AC6D-492E-B098-190166F0C44B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServerTests", "ServerTests\ServerTests.csproj", "{62494D85-1181-40AF-88E1-CD621507F79F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -38,6 +40,10 @@ Global
 		{E49D61BD-AC6D-492E-B098-190166F0C44B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E49D61BD-AC6D-492E-B098-190166F0C44B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E49D61BD-AC6D-492E-B098-190166F0C44B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{62494D85-1181-40AF-88E1-CD621507F79F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{62494D85-1181-40AF-88E1-CD621507F79F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{62494D85-1181-40AF-88E1-CD621507F79F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{62494D85-1181-40AF-88E1-CD621507F79F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
